Make training a real subpackage instead of a build-time copy
force-include shipped the scripts by copying training/ into the wheel at build
time. That is duplication by another name: the same files at two paths, imports
resolving by accident of sys.path, and nothing type-checking any of it.

training/ is now src/piedomains/training/, a genuine subpackage. Ten
sys.path.insert hacks deleted, thirteen cross-script imports made relative, and
every script runs as python -m piedomains.training.<name> -- they all already had
__main__ guards. __init__.py imports nothing, so import piedomains.training stays
at 17ms and does not drag torch in for callers who only want to list the scripts.

Bringing them under src/ put them under pyright and pydoclint for the first time,
which immediately found a break I had already made: fuse.py had two
function-local imports (from train_text import TextDataset, from train_image
import ScreenshotDataset) that the module-level pass missed. Without sys.path
those fail at runtime, so fusion would have broken on install. Twenty-two other
errors fixed with it -- six __doc__.split() calls that are None under python -OO,
five DataLoader arguments where torch's stub is stricter than its runtime
contract, and a function shadowed by a same-named variable.

tests/test_cli.py gains a check that no sys.path.insert survives anywhere in the
subpackage, because that is exactly the pattern that resolved when run as a
script from the right directory and failed as an import.

Verified by building a wheel, installing into a clean venv, and asserting the
scripts resolve inside site-packages and import as modules.
