+++ pyproject.toml
@@ -88,15 +88,6 @@ exclude = [
 [tool.hatch.build.targets.wheel]
 packages = ["src/piedomains"]
 
-# The training and evaluation scripts ship with the package. Every accuracy claim in
-# the README comes out of these, and a claim you cannot re-run is a claim you have to
-# take on faith -- `pip install piedomains` should be enough to reproduce them.
-#
-# force-include rather than a src/piedomains/training package because the scripts import
-# each other by path (`sys.path.insert(0, Path(__file__).parent)`), which keeps them
-# runnable straight from a checkout as well as from site-packages.
-[tool.hatch.build.targets.wheel.force-include]
-"training" = "piedomains/training"
 
 # Ruff configuration
 
@@ -128,8 +119,8 @@ convention = "google"
 # `docker run --tmpfs` arguments rather than paths this process writes to.
 "examples/**" = ["T201", "D", "S101", "S603", "S607", "S108", "RUF059"]
 "streamlit/**" = ["T201", "D"]
-# Training/eval scripts are CLIs: printing the report is their whole purpose.
-"training/**" = ["T201"]
+# Training/eval scripts are CLIs: reporting progress to stdout is their whole purpose.
+"src/piedomains/training/**" = ["T201"]
 
 [tool.pytest.ini_options]
 minversion = "7.0"
